4_kms_dek/requirements.txt
google-cloud-pubsub
canonicaljson
PyJWT
google-api-python-client
httplib2
oauth2client
requests
google-auth-httplib2
cryptography
expiringdict
